Refactor: remove unused code from PersistenceManager
PersistenceManager had a lot of definition that wasn't used. It either had no implementation, or the little implementation that was there, went unreferenced. This commit removes it all, reducing code complexity by a significant factor.

It appears to me that all removed code was replaced by code in the implementation of ConversationManager - although I'm not 100% sure.